zk/acc_input_hash/acc_input_hash.go
package acc_input_hash

import (
	"context"
	"fmt"

	"github.com/ledgerwatch/erigon-lib/chain"
	"github.com/ledgerwatch/erigon-lib/common"
	"github.com/ledgerwatch/erigon-lib/kv"
	"github.com/ledgerwatch/erigon/core/state"
	"github.com/ledgerwatch/erigon/core/systemcontracts"
	eritypes "github.com/ledgerwatch/erigon/core/types"
	"github.com/ledgerwatch/erigon/zk/utils"
)

const SpecialZeroHash = "0x27AE5BA08D7291C96C8CBDDCC148BF48A6D68C7974B94356F53754EF6171D757"

// BlockReader is a rawdb block reader abstraction for easier testing
type BlockReader interface {
	ReadBlockByNumber(blockNo uint64) (*eritypes.Block, error)
}

// BatchDataReader is an abstraction for reading batch data
type BatchDataReader interface {
	GetEffectiveGasPricePercentage(txHash common.Hash) (uint8, error)
	GetL2BlockNosByBatch(batchNo uint64) ([]uint64, error)
	GetForkId(batchNo uint64) (uint64, error)
}

// L1DataReader is an abstraction for reading L1 data
type L1DataReader interface {
	GetL1InfoTreeIndexToRoots() (map[uint64]common.Hash, error)
	GetBlockL1InfoTreeIndex(blockNo uint64) (uint64, error)
}

// AccInputHashReader combines the necessary reader interfaces
type AccInputHashReader interface {
	GetAccInputHashForBatchOrPrevious(batchNo uint64) (common.Hash, uint64, error)
	BatchDataReader
	L1DataReader
}

// AccInputHashCalculator is an abstraction for calculating acc input hashes
type AccInputHashCalculator interface {
	Calculate(batchNum uint64) (common.Hash, error)
}

type BaseCalc struct {
	Ctx         context.Context
	Reader      AccInputHashReader
	Tx          kv.Tx
	BlockReader BlockReader
}

var Calculators = map[uint64]func(*BaseCalc) AccInputHashCalculator{
	// pre 7 not supported
	7:  NewFork7Calculator, // etrog
	8:  NewFork7Calculator,
	9:  NewFork7Calculator,
	10: NewFork7Calculator,
	11: NewFork7Calculator,
	12: NewFork7Calculator,
	13: NewFork7Calculator,
}

func NewCalculator(ctx context.Context, tx kv.Tx, reader AccInputHashReader, forkID uint64) (AccInputHashCalculator, error) {
	calcConstructor, ok := Calculators[forkID]
	if !ok {
		return nil, fmt.Errorf("unsupported fork ID: %d", forkID)
	}

	baseCalc := &BaseCalc{
		Ctx:         ctx,
		Reader:      reader,
		Tx:          tx,
		BlockReader: &RawDbBlockReader{tx: tx},
	}

	return calcConstructor(baseCalc), nil
}

func NewCalculatorWithBlockReader(ctx context.Context, tx kv.Tx, reader AccInputHashReader, blockReader BlockReader, forkID uint64) (AccInputHashCalculator, error) {
	calcConstructor, ok := Calculators[forkID]
	if !ok {
		return nil, fmt.Errorf("unsupported fork ID: %d", forkID)
	}

	baseCalc := &BaseCalc{
		Ctx:         ctx,
		Reader:      reader,
		Tx:          tx,
		BlockReader: blockReader,
	}

	return calcConstructor(baseCalc), nil
}

type PreFork7Calculator struct {
	*BaseCalc
}

func NewPreFork7Calculator(bc *BaseCalc) AccInputHashCalculator {
	return PreFork7Calculator{BaseCalc: bc}
}

func (p PreFork7Calculator) Calculate(batchNum uint64) (common.Hash, error) {
	// this isn't supported
	return common.Hash{}, nil
}

type Fork7Calculator struct {
	*BaseCalc
}

func NewFork7Calculator(bc *BaseCalc) AccInputHashCalculator {
	return Fork7Calculator{BaseCalc: bc}
}

func (f Fork7Calculator) Calculate(batchNum uint64) (common.Hash, error) {
	accInputHash, returnedBatchNo, err := f.Reader.GetAccInputHashForBatchOrPrevious(batchNum)
	if err != nil {
		return common.Hash{}, err
	}

	// check the forkid of the returnedBatchNo
	forkId, err := f.Reader.GetForkId(returnedBatchNo)
	if err != nil {
		return common.Hash{}, err
	}

	if forkId < uint64(chain.ForkID7Etrog) {
		return common.Hash{}, fmt.Errorf("unsupported fork ID: %d", forkId)
	}

	// TODO: remove test spoooofing! (1001 and 997 are l1 held batch accinputhashes - sequence ends)
	if batchNum >= 997 {
		// let's just spoof it backwards:
		accInputHash, returnedBatchNo, err = f.Reader.GetAccInputHashForBatchOrPrevious(995)
		if err != nil {
			return common.Hash{}, err
		}
	}

	// if we have it, return it
	if returnedBatchNo == batchNum {
		return accInputHash, nil
	}

	// otherwise calculate it
	accInputHash, err = LocalAccInputHashCalc(f.Ctx, f.Reader, f.BlockReader, f.Tx, batchNum, returnedBatchNo, accInputHash)
	if err != nil {
		return common.Hash{}, err
	}

	return accInputHash, nil
}

func LocalAccInputHashCalc(ctx context.Context, reader AccInputHashReader, blockReader BlockReader, tx kv.Tx, batchNum, startBatchNo uint64, prevAccInputHash common.Hash) (common.Hash, error) {
	var accInputHash common.Hash

	infoTreeIndexes, err := reader.GetL1InfoTreeIndexToRoots()
	if err != nil {
		return common.Hash{}, fmt.Errorf("failed to get l1 info tree indexes: %w", err)
	}
	if len(infoTreeIndexes) == 0 {
		return common.Hash{}, fmt.Errorf("no l1 info tree indexes found")
	}

	// go from injected batch with known batch 0 accinputhash of 0x0...0
	if startBatchNo == 0 {
		startBatchNo = 1
	}

	// TODO: handle batch 1 case where we should get check the aggregator code: https://github.com/0xPolygon/cdk/blob/develop/aggregator/aggregator.go#L1167

	for i := startBatchNo; i <= batchNum; i++ {
		currentForkId, err := reader.GetForkId(i)
		if err != nil {
			return common.Hash{}, fmt.Errorf("failed to get fork id for batch %d: %w", i, err)
		}

		batchBlockNos, err := reader.GetL2BlockNosByBatch(i)
		if err != nil {
			return common.Hash{}, fmt.Errorf("failed to get batch blocks for batch %d: %w", i, err)
		}
		batchBlocks := []*eritypes.Block{}
		var coinbase common.Address
		for in, blockNo := range batchBlockNos {
			block, err := blockReader.ReadBlockByNumber(blockNo)
			if err != nil {
				return common.Hash{}, fmt.Errorf("failed to get block %d: %w", blockNo, err)
			}
			if in == 0 {
				coinbase = block.Coinbase()
			}
			batchBlocks = append(batchBlocks, block)
		}

		lastBlockNoInPreviousBatch := uint64(0)
		firstBlockInBatch := batchBlocks[0]
		if firstBlockInBatch.NumberU64() != 0 {
			lastBlockNoInPreviousBatch = firstBlockInBatch.NumberU64() - 1
		}

		lastBlockInPreviousBatch, err := blockReader.ReadBlockByNumber(lastBlockNoInPreviousBatch)
		if err != nil {
			return common.Hash{}, err
		}

		batchL2Data, err := utils.GenerateBatchDataFromDb(tx, reader, batchBlocks, lastBlockInPreviousBatch, currentForkId)
		if err != nil {
			return common.Hash{}, fmt.Errorf("failed to generate batch data for batch %d: %w", i, err)
		}

		highestBlock := batchBlocks[len(batchBlocks)-1]

		sr := state.NewPlainState(tx, highestBlock.NumberU64(), systemcontracts.SystemContractCodeLookup["hermez"])
		if err != nil {
			return common.Hash{}, fmt.Errorf("failed to get psr: %w", err)
		}
		l1InfoRootBytes, err := sr.ReadAccountStorage(state.ADDRESS_SCALABLE_L2, 1, &state.BLOCK_INFO_ROOT_STORAGE_POS)
		if err != nil {
			return common.Hash{}, fmt.Errorf("failed to read l1 info root: %w", err)
		}
		sr.Close()
		l1InfoRoot := common.BytesToHash(l1InfoRootBytes)

		limitTs := highestBlock.Time()

		fmt.Println("[l1InfoRoot]", l1InfoRoot.Hex())
		fmt.Println("[limitTs]", limitTs)

		inputs := utils.AccHashInputs{
			OldAccInputHash: prevAccInputHash,
			Sequencer:       coinbase,
			BatchData:       batchL2Data,
			L1InfoRoot:      l1InfoRoot,
			LimitTimestamp:  limitTs,
			ForcedBlockHash: common.Hash{},
		}
		accInputHash, err = utils.CalculateAccInputHashByForkId(inputs)
		if err != nil {
			return common.Hash{}, fmt.Errorf("failed to calculate accInputHash for batch %d: %w", i, err)
		}
		prevAccInputHash = accInputHash
	}
	return accInputHash, nil
}
